requirements.txt
streamlit
playwright
google-genai
nest_asyncio